+++ src/app/component/visit-summary/patient-interaction/patient-interaction.component.ts
@@ -11,9 +11,7 @@ import {
   keyframes,
 } from "@angular/animations";
 import { DiagnosisService } from "src/app/services/diagnosis.service";
-import { TranslationService } from "src/app/services/translation.service";
-declare var getEncounterProviderUUID: any,
-  getFromStorage: any,
+declare var getFromStorage: any,
   getEncounterUUID: any;
 
 @Component({
@@ -70,7 +68,6 @@ export class PatientInteractionComponent implements OnInit {
   constructor(
     private diagnosisService: DiagnosisService,
     private visitService: VisitService,
-    private translationService: TranslationService,
     private route: ActivatedRoute,
     private encounterService: EncounterService
   ) {}
@@ -142,8 +139,7 @@ export class PatientInteractionComponent implements OnInit {
     const formValue = this.interaction.value;
     const value = formValue.interaction;
     const providerDetails = getFromStorage("provider");
-    const providerUuid = providerDetails.uuid;
-    if (providerDetails && providerUuid === getEncounterProviderUUID()) {
+    if (this.diagnosisService.isSameDoctor()) {
       this.visitService.getAttribute(visitId).subscribe((response) => {
         const result = response.results;
         if (result.length !== 0 && ["Yes", "No"].includes(response.value)) {
@@ -194,19 +190,19 @@ export class PatientInteractionComponent implements OnInit {
           });
         }
       }
-    } else {
-      this.translationService.getTranslation("Another doctor is viewing this case");
     }
   }
 
   delete(i) {
-    this.visitService.deleteAttribute(this.visitId, i).subscribe((res) => {
-      this.msg = [];
-    });
-    if (this.adviceObs.length > 0) {
-      this.adviceObs.forEach(({ uuid }) => {
-        this.diagnosisService.deleteObs(uuid).subscribe();
+    if (this.diagnosisService.isSameDoctor()) {
+      this.visitService.deleteAttribute(this.visitId, i).subscribe((res) => {
+        this.msg = [];
       });
+      if (this.adviceObs.length > 0) {
+        this.adviceObs.forEach(({ uuid }) => {
+          this.diagnosisService.deleteObs(uuid).subscribe();
+        });
+      }
     }
   }
 }
